+++ src/oldnews/commands/main.py
@@ -77,4 +77,32 @@ class MarkAllRead(Command):
     BINDING_KEY = "R"
 
 
+##############################################################################
+class CopyHomePageToClipboard(Command):
+    """Copy the URL of the current subscription's home page to the clipboard"""
+
+    BINDING_KEY = "f3"
+
+
+##############################################################################
+class CopyFeedToClipboard(Command):
+    """Copy the URL of the current subscription's feed to the clipboard"""
+
+    BINDING_KEY = "f4"
+
+
+##############################################################################
+class CopyArticleToClipboard(Command):
+    """Copy the URL for the current article to the clipboard"""
+
+    BINDING_KEY = "f5"
+
+
+##############################################################################
+class Copy(Command):
+    """Copy a URL to the clipboard depending on the context"""
+
+    BINDING_KEY = "ctrl+c"
+
+
 ### main.py ends here