+++ src/oldnews/providers/main.py
@@ -13,6 +13,10 @@ from textual_enhanced.commands import (
 ##############################################################################
 # Local imports.
 from ..commands import (
+    Copy,
+    CopyArticleToClipboard,
+    CopyFeedToClipboard,
+    CopyHomePageToClipboard,
     Escape,
     MarkAllRead,
     Next,
@@ -44,6 +48,10 @@ class MainCommands(CommandsProvider):
         yield from self.maybe(OpenArticle)
         yield from self.maybe(OpenHomePage)
         yield from self.maybe(MarkAllRead)
+        yield from self.maybe(CopyHomePageToClipboard)
+        yield from self.maybe(CopyFeedToClipboard)
+        yield from self.maybe(CopyArticleToClipboard)
+        yield from self.maybe(Copy)
         yield ToggleShowAll()
         yield RefreshFromTheOldReader()
         yield ChangeTheme()